[v1.0.13] - separated <Modal /> as a module
