+++ src/components/common/Modal/index.js
@@ -1,0 +1,2 @@
+import Modal from './Modal'
+export default Modal
