+++ src/components/common/button/EditButtonSet/EditButtonSet.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 
 import Button from '@material-ui/core/Button'
-import Modal from '@material-ui/core/Modal'
+import Modal from '../../Modal'
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -24,41 +24,10 @@ const useStyles = makeStyles(theme => ({
   cancel: {
 
   },
-  modalPaper: {
-    position: 'absolute',
-    width: 400,
-    backgroundColor: '#ffffff',
-    border: '1px solid rgba(0, 0, 0, 0.23)',
-    borderRadius: '5px',
-    boxShadow: theme.shadows[5],
-    padding: theme.spacing(2,4,3),
-    outline: 0,
-
-    '& h2': {
-      fontWeight: 300,
-    },
-
-    '& button': {
-      marginRight: '3px',
-      marginLeft: '3px',
-    }
-  },
 }))
-
-const getModalStyle = () => {
-  const top = 50
-  const left = 50
-
-  return {
-    top: `${top}%`,
-    left: `${left}%`,
-    transform: `translate(-${top}%, -${left}%)`,
-  };
-}
 
 const EditButtonSet = (props) => {
   const classes = useStyles()
-  const [modalStyle] = useState(getModalStyle)
   const [updateModalOpen, setUpdateModalOpen] = useState(false)
   const [deleteModalOpen, setDeleteModalOpen] = useState(false)
 
@@ -83,12 +52,10 @@ const EditButtonSet = (props) => {
         수정
       </Button>
       <Modal
-        aria-labelledby="simple-modal-title"
-        open={updateModalOpen}
+        modalOpen={updateModalOpen}
+        title="이대로 수정하시겠어요?"
       >
-        <div style={modalStyle} className={classes.modalPaper}>
-          <h2>이대로 수정하시겠어요?</h2>
-          <Button
+        <Button
             variant="outlined"
             className={classes.update}
             onClick={props.onClickUpdate}
@@ -102,7 +69,6 @@ const EditButtonSet = (props) => {
           >
             아니오
           </Button>
-        </div>
       </Modal>
       <Button
         variant="outlined"
@@ -112,12 +78,10 @@ const EditButtonSet = (props) => {
         삭제
       </Button>
       <Modal
-        aria-labelledby="simple-modal-title"
-        open={deleteModalOpen}
+        modalOpen={deleteModalOpen}
+        title="정말로 삭제하시겠어요?"
       >
-        <div style={modalStyle} className={classes.modalPaper}>
-          <h2>정말로 삭제하시겠어요?</h2>
-          <Button
+        <Button
             variant="outlined"
             className={classes.delete}
             onClick={props.onClickDelete}
@@ -131,7 +95,6 @@ const EditButtonSet = (props) => {
           >
             아니오
           </Button>
-        </div>
       </Modal>
       <Button
         variant="outlined"
